Add version 0.9

+++ app/Http/Controllers/CompanyPosController.php
@@ -6,6 +6,11 @@ use App\Models\CompanyPos;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\DB;
+use App\Models\Service;
+use App\Models\Provider;
+use App\Models\Staff;
+use App\Models\OtherPay;
 
 class CompanyPosController extends Controller
 {
@@ -14,21 +19,28 @@ class CompanyPosController extends Controller
      */
     public function index(): View
     {
-        return view('companypos.index');
+        $list = Staff::all();
+        return view('companypos.index', ['list' => $list]);
     }
 
     public function list(Request $request): JsonResponse
     {
-        $list = CompanyPos::all();
+        $list = CompanyPos::select('companyPos.id', 'companyPos.pos', 'companyPos.commission', 'companyPos.contactName', 'companyPos.contactPhone', 'companyPos.description', 
+            'staff.name as staffName', 'staff.id as staffId', 'ruc', 'bank', 'accountNumber')
+            ->leftJoin('staff', 'companyPos.staffId', '=', 'staff.id')
+            ->get();
         return response()->json(['list' => $list]);
     }
 
     public function add(Request $request)
     {
-        //todo: verificar que ya existe el ruc y el dni
         $companyPos = new CompanyPos();
         $companyPos->pos = $request->pos;
         $companyPos->commission = $request->commission;
+        $companyPos->staffId = $request->staffId;
+        $companyPos->ruc = $request->ruc;
+        $companyPos->bank = $request->bank;
+        $companyPos->accountNumber = $request->accountNumber; 
         $companyPos->contactName = $request->contactName;
         $companyPos->contactPhone = $request->contactPhone;
         $companyPos->description = $request->description;
@@ -42,6 +54,10 @@ class CompanyPosController extends Controller
         $companyPos = CompanyPos::find($request->companyPosId);
         $companyPos->pos = $request->pos;
         $companyPos->commission = $request->commission;
+        $companyPos->staffId = $request->staffId;
+        $companyPos->ruc = $request->ruc;
+        $companyPos->bank = $request->bank;
+        $companyPos->accountNumber = $request->accountNumber; 
         $companyPos->contactName = $request->contactName;
         $companyPos->contactPhone = $request->contactPhone;
         $companyPos->description = $request->description;
@@ -52,15 +68,28 @@ class CompanyPosController extends Controller
 
     public function remove(Request $request): JsonResponse
     {
-        //todo: verificar que no haya ventas asociadas al pos
-        CompanyPos::find($request->companyPosId)->delete();      
-        return response()->json(['status'=>'success', 'message'=>'El POS fue eliminado']);     
+        $rows = DB::table('sales')->where('companyPosId', $request->companyPosId)->count();
+        if($rows == 0) {
+            CompanyPos::find($request->companyPosId)->delete();      
+            return response()->json(['status'=>'success', 'message'=>'El POS fue eliminado']);         
+        }else{
+            return response()->json(['status'=>'error', 'message'=>'No se puede eliminar un POS con ventas relacionadas']);     
+        }   
+    }
 
-        // $rows = DB::table('sales')->where('clientId', $request->clientId)->count();
-        // if($rows == 0) {
-            
-        // }else{
-        //     return response()->json(['status'=>'error', 'message'=>'No se puede eliminar un cliente con ventas relacionadas']);     
-        // }   
+
+    public function detail(Request $request): View
+    {
+        $companyPos = CompanyPos::find($request->companyPosId);
+
+        $services = Service::all();
+
+        $providers = Provider::all();
+
+        $staffs = Staff::all();
+
+        $otherpays = OtherPay::all();
+
+        return view('companypos.detail', ['companyPos' => $companyPos, 'services' => $services, 'staffs' => $staffs, 'providers' => $providers, 'otherpays' => $otherpays]);
     }
 }
